+++ hpc/nvidia/cuda/fakeCUDA/nccl.hpp
@@ -7,6 +7,8 @@ extern "C" {
 #endif
 
 ncclResult_t ncclGetVersion(int *version);
+const char* ncclGetErrorString(ncclResult_t result);
+ncclResult_t ncclGetUniqueId(ncclUniqueId* uniqueId);
 
 #ifdef __cplusplus
 }